Slot everything into place and remove old code.
