+++ webapp/src/ts/services/enketo.service.ts
@@ -6,18 +6,22 @@ import * as FileManager from '../../js/enketo/file-manager.js';
 import events from 'enketo-core/src/js/event';
 
 import { Xpath } from '@mm-providers/xpath-element-path.provider';
-import { AttachmentService } from '@mm-services/attachment.service';
 import { DbService } from '@mm-services/db.service';
 import { EnketoPrepopulationDataService } from '@mm-services/enketo-prepopulation-data.service';
-import { EnketoTranslationService } from '@mm-services/enketo-translation.service';
 import { ExtractLineageService } from '@mm-services/extract-lineage.service';
-import { REPORT_ATTACHMENT_NAME } from '@mm-services/get-report-content.service';
 import { TranslateFromService } from '@mm-services/translate-from.service';
 import { TranslateService } from '@mm-services/translate.service';
 import { CHTDatasourceService } from '@mm-services/cht-datasource.service';
-import { Qualifier, Report } from '@medic/cht-datasource';
+import { Contact, Qualifier } from '@medic/cht-datasource';
 import { DOC_TYPES } from '@medic/constants';
-import { EnketoForm } from '@mm-services/NewEnketoService';
+import { FormConfig } from '@mm-services/form/form-config';
+import {
+  EnektoContactFormData,
+  EnketoFormData,
+  EnketoReportFormData,
+  EnketoRootFormData
+} from '@mm-services/form/form-data';
+import { isHardcodedType } from '@medic/contact-types-utils';
 
 /**
  * Service for interacting with Enketo forms. This code is intended for displaying forms in the CHT as well as being
@@ -30,21 +34,22 @@ import { EnketoForm } from '@mm-services/NewEnketoService';
 })
 export class EnketoService {
   constructor(
-    private attachmentService: AttachmentService,
     private dbService: DbService,
     private enketoPrepopulationDataService: EnketoPrepopulationDataService,
-    private enketoTranslationService: EnketoTranslationService,
     private extractLineageService: ExtractLineageService,
     private translateFromService: TranslateFromService,
     private translateService: TranslateService,
     chtDatasourceService: CHTDatasourceService,
     private ngZone: NgZone,
   ) {
-    this.getReport = chtDatasourceService.bind(Report.v1.get);
+    this.getContactFromDatasource = chtDatasourceService.bind(Contact.v1.get);
   }
 
+  private readonly USER_BINARY_ATTACHMENT_PREFIX = 'user-file';
+  private readonly USER_FILE_ATTACHMENT_PREFIX = `${this.USER_BINARY_ATTACHMENT_PREFIX}-`;
+
   private readonly objUrls: string[] = [];
-  private readonly getReport: ReturnType<typeof Report.v1.get>;
+  private readonly getContactFromDatasource: ReturnType<typeof Contact.v1.get>;
   private currentForm;
 
   getCurrentForm() {
@@ -364,181 +369,6 @@ export class EnketoService {
     return { form, config: formConfig };
   }
 
-  private xmlToDocs(doc, formXml, xmlVersion, record) {
-    const recordDoc = $.parseXML(record);
-    const $record = $($(recordDoc).children()[0]);
-    const repeatPaths = this.enketoTranslationService.getRepeatPaths(formXml);
-
-    const mapOrAssignId = (e, id?) => {
-      if (!id) {
-        const $id = $(e).children('_id');
-        if ($id.length) {
-          id = $id.text();
-        }
-        if (!id) {
-          id = uuid();
-        }
-      }
-      e._couchId = id;
-    };
-
-    mapOrAssignId($record[0], doc._id || uuid());
-
-    const getId = (xpath) => {
-      const xPathResult = recordDoc.evaluate(xpath, recordDoc, null, XPathResult.ANY_TYPE, null);
-      let node = xPathResult.iterateNext();
-      while (node) {
-        if (node._couchId) {
-          return node._couchId;
-        }
-        node = xPathResult.iterateNext();
-      }
-    };
-
-    const getRelativePath = (path) => {
-      if (!path) {
-        return;
-      }
-      path = path.trim();
-
-      const repeatReference = repeatPaths?.find(repeatPath => path === repeatPath || path.startsWith(`${repeatPath}/`));
-      if (repeatReference) {
-        if (repeatReference === path) {
-          // when the path is the repeat element itself, return the repeat element node name
-          return path.split('/').slice(-1)[0];
-        }
-
-        return path.replace(`${repeatReference}/`, '');
-      }
-
-      if (path.startsWith('./')) {
-        return path.replace('./', '');
-      }
-    };
-
-    const getClosestPath = (element, $element, path) => {
-      const relativePath = getRelativePath(path);
-      if (!relativePath) {
-        return;
-      }
-
-      // assign a unique id for xpath context, since the element can be inside a repeat
-      if (!element.id) {
-        element.id = uuid();
-      }
-      const uniqueElementSelector = `${element.nodeName}[@id="${element.id}"]`;
-
-      const closestPath = `//${uniqueElementSelector}/ancestor-or-self::*/descendant-or-self::${relativePath}`;
-      try {
-        recordDoc.evaluate(closestPath, recordDoc);
-        return closestPath;
-      } catch (err) {
-        console.error('Error while evaluating closest path', closestPath, err);
-      }
-    };
-
-    // Chrome 30 doesn't support $xml.outerHTML: #3880
-    const getOuterHTML = (xml) => {
-      if (xml.outerHTML) {
-        return xml.outerHTML;
-      }
-      return $('<temproot>').append($(xml).clone()).html();
-    };
-
-    const dbDocTags: string[] = [];
-    $record
-      .find('[db-doc]')
-      .filter((idx, element) => {
-        return $(element).attr('db-doc')?.toLowerCase() === 'true';
-      })
-      .each((idx, element) => {
-        mapOrAssignId(element);
-        dbDocTags.push(element.tagName);
-      });
-
-    $record
-      .find('[db-doc-ref]')
-      .each((idx, element) => {
-        const $element = $(element);
-        const reference = $element.attr('db-doc-ref');
-        const path = getClosestPath(element, $element, reference);
-
-        const refId = (path && getId(path)) || getId(reference);
-        $element.text(refId);
-      });
-
-    const docsToStore = $record
-      .find('[db-doc=true]')
-      .map((idx, element) => {
-        const docToStore: any = this.enketoTranslationService.reportRecordToJs(getOuterHTML(element));
-        docToStore._id = getId(Xpath.getElementXPath(element));
-        docToStore.reported_date = Date.now();
-        return docToStore;
-      })
-      .get();
-
-    doc._id = getId('/*');
-    if (xmlVersion) {
-      doc.form_version = xmlVersion;
-    }
-    doc.hidden_fields = this.enketoTranslationService.getHiddenFieldList(record, dbDocTags);
-
-    FileManager
-      .getCurrentFiles()
-      .forEach(file => this.attachmentService.add(doc, `user-file-${file.name}`, file, file.type, false));
-
-    const attachLegacyFile = (elem, file, type, alreadyEncoded) => {
-      const xpath = Xpath.getElementXPath(elem);
-      // replace instance root element node name with form internal ID
-      const filename = 'user-file' +
-        (xpath.startsWith('/' + doc.form) ? xpath : xpath.replace(/^\/[^/]+/, '/' + doc.form));
-      this.attachmentService.add(doc, filename, file, type, alreadyEncoded);
-    };
-
-    $record
-      .find('[type=binary]')
-      .each((idx, element) => {
-        const file = $(element).text();
-        if (file) {
-          // Attach binary file with legacy-style filename because the actual filename is not stored as the question
-          // value in the form model (and so there is currently no way to map the answer in a saved report to the
-          // associated file attachment).
-          attachLegacyFile(element, file, 'image/png', true);
-        }
-      });
-
-    record = getOuterHTML($record[0]);
-
-    // remove old style content attachment
-    this.attachmentService.remove(doc, REPORT_ATTACHMENT_NAME);
-    docsToStore.unshift(doc);
-
-    doc.fields = this.enketoTranslationService.reportRecordToJs(record, formXml);
-    return docsToStore;
-  }
-
-  private async update(docId) {
-    // update an existing doc.  For convenience, get the latest version
-    // and then modify the content.  This will avoid most concurrent
-    // edits, but is not ideal.
-    const doc = await this.getReport(Qualifier.byUuid(docId));
-    // previously XML was stored in the content property
-    // TODO delete this and other "legacy" code support commited against
-    //      the same git commit at some point in the future?
-    return { ...doc, content: undefined };
-  }
-
-  private create(formInternalId, contact) {
-    return {
-      form: formInternalId,
-      type: DOC_TYPES.DATA_RECORD,
-      content_type: 'xml',
-      reported_date: Date.now(),
-      contact: this.extractLineageService.extract(contact),
-      from: contact?.phone
-    };
-  }
-
   private forceRecalculate(form) {
     // Work-around for stale jr:choice-name() references in labels.  ref #3870
     form.calc.update();
@@ -566,33 +396,221 @@ export class EnketoService {
     }
   }
 
-  private async prepareForSave(form) {
+  public async saveContact({ config, form }: EnketoForm, defaultData: Record<string, any>) {
+    return this.ngZone.runOutsideAngular(async () => {
+      await this.validate(form);
+      const contactDoc = this.initializeDoc(defaultData);
+      const formData = new EnektoContactFormData(
+        this.getFormDataXml(form),
+        contactDoc._id,
+        isHardcodedType(contactDoc.type) ? contactDoc.type : contactDoc.contact_type
+      );
+
+      const formAttachments = this.processFormAttachments(config.doc.internalId, formData, contactDoc._attachments);
+
+      const rootOutputDoc: Record<string, any> = {
+        ...contactDoc,
+        ...formData.deserializeDoc(config),
+        type: contactDoc.type,
+        contact_type: contactDoc.contact_type,
+        _attachments: formAttachments
+      };
+
+      const siblings = EnektoContactFormData.SIBLING_FIELD_NAMES
+        .map(fieldName => ({ fieldName, doc: formData.getSiblingData(fieldName)?.deserializeDoc(config) }))
+        .map(({ fieldName, doc }) => ({ fieldName, doc: this.initializeContactSibling(rootOutputDoc, doc)}));
+      await Promise.all(siblings.map(async ({ fieldName, doc }) => {
+        rootOutputDoc[fieldName] = await this.getContactSiblingValue(
+          doc, rootOutputDoc[fieldName], defaultData[fieldName]
+        );
+      }));
+      const outputSiblings = siblings
+        .filter(({ fieldName, doc }) => doc && rootOutputDoc[fieldName] === doc)
+        .map(({ doc }) => doc)
+        .filter(doc => !!doc);
+
+      const childDocs = formData
+        .getChildData()
+        .map(doc => this.initializeDoc(doc.deserializeDoc(config)))
+        .map(doc => ({ ...doc, parent: rootOutputDoc }));
+
+      return {
+        docId: rootOutputDoc._id,
+        preparedDocs: [rootOutputDoc, ...outputSiblings, ...childDocs].map(doc => this.dehydrateContactLineage(doc))
+      };
+    });
+  }
+
+  public async saveReport({ config, form }: EnketoForm, defaultData: Record<string, any>) {
+    return this.ngZone.runOutsideAngular(async () => {
+      await this.validate(form);
+      const { internalId, xmlVersion } = config.doc;
+      const reportDoc: Record<string, any> = this.initializeReportDoc(internalId, xmlVersion, defaultData);
+      const formData = new EnketoReportFormData(this.getFormDataXml(form), reportDoc._id);
+      const subDocsData = formData.getDbDocData();
+
+      this.populateDbDocRefElements(formData, [formData, ...subDocsData]);
+      const attachments = this.processFormAttachments(config.doc.internalId, formData, reportDoc._attachments);
+      const hiddenFields = this.getHiddenFields([
+        ...formData.hiddenElements,
+        ...subDocsData.map(({ rootElement }) => rootElement)
+      ]);
+
+      const rootOutputDoc: Record<string, any> = {
+        ...reportDoc,
+        hidden_fields: hiddenFields,
+        fields: formData.deserialize(config),
+        _attachments: attachments
+      };
+
+      const dbDocObjects = subDocsData
+        .map(docData => docData.deserializeDoc(config))
+        .map(doc => this.initializeDoc(doc));
+      return [rootOutputDoc, ...dbDocObjects];
+    });
+  }
+
+  private async validate(form: Record<string, any>) {
     const valid = await form.validate();
     if (!valid) {
-      throw new Error('Form is invalid');
+      throw new FormValidationError();
     }
     form.view.html.dispatchEvent(events.BeforeSave());
   }
 
-  async completeNewReport(formInternalId, form, formDoc, contact) {
-    await this.prepareForSave(form);
-    return this.ngZone.runOutsideAngular(async () => {
-      const doc = this.create(formInternalId, contact);
-      return this._save(form, formDoc, doc);
+  private getFormDataXml(form: Record<string, any>) {
+    const formString = form.getDataStr({ irrelevant: false });
+    return new DOMParser().parseFromString(formString, 'text/xml');
+  }
+
+  private initializeContactSibling(rootContactDoc: Record<string, any>, rawSibling?: Record<string, any>) {
+    if (!rawSibling) {
+      return;
+    }
+    return {
+      type: 'person', // legacy support - form data should override this
+      ...this.initializeDoc(rawSibling),
+      parent: rawSibling.parent === 'PARENT' ? { ...rootContactDoc } : rawSibling.parent
+    };
+  }
+
+  private async getContactSiblingValue(
+    sibling?: Record<string, any>,
+    currentValue?: Record<string, string>,
+    defaultValue?: Record<string, unknown>
+  ) {
+    if (!currentValue?._id) {
+      return undefined;
+    }
+    if (currentValue._id === 'NEW') {
+      return sibling;
+    }
+    if (currentValue._id === defaultValue?._id) {
+      return defaultValue;
+    }
+
+    return await this.getContactFromDatasource(Qualifier.byUuid(currentValue._id));
+  }
+
+  private dehydrateContactLineage(contactDoc: Record<string, any>) {
+    return {
+      ...contactDoc,
+      parent: this.extractLineageService.extract(contactDoc.parent),
+      contact: this.extractLineageService.extract(contactDoc.contact)
+    };
+  }
+
+  private getHiddenFields(elements: Element[]) {
+    const hiddenXpaths = new Set(elements.map((element) => Xpath.getElementRawXPath(element)));
+    const hasHiddenAncestor = (
+      segments: string[]
+    ) => (_: string, i: number) => i > 0 && hiddenXpaths.has(segments.slice(0, i).join('/'));
+    return [...hiddenXpaths]
+      .map(xpath => xpath.split('/'))
+      .filter(segments => !segments.some(hasHiddenAncestor(segments)))
+      .map(segments => segments
+        .filter(Boolean)
+        .slice(1)
+        .join('.'));
+  }
+
+  private initializeDoc(defaultData: Record<string, any>): Record<string, any> {
+    return {
+      _id: defaultData._id || uuid(),
+      reported_date: Date.now(),
+      ...defaultData
+    };
+  }
+
+  private initializeReportDoc(form: string, formVersion: string, defaultData: Record<string, any>) {
+    return {
+      form,
+      type: DOC_TYPES.DATA_RECORD,
+      content_type: 'xml',
+      from: defaultData.contact?.phone,
+      ...this.initializeDoc(defaultData),
+      contact: this.extractLineageService.extract(defaultData.contact),
+      form_version: formVersion
+    };
+  }
+
+  private populateDbDocRefElements(formData: EnketoReportFormData, allData: EnketoFormData[]) {
+    formData.dbDocRefElements.forEach(element => {
+      const reference = element.getAttribute('db-doc-ref');
+      const referencedNode = formData.getNodeByXpath(element, reference);
+      if (!referencedNode) {
+        return;
+      }
+      const refDoc = allData.find(({ rootElement }) => rootElement === referencedNode);
+      if (refDoc) {
+        element.textContent = refDoc.id;
+      }
     });
   }
 
-  async completeExistingReport(form, formDoc, docId) {
-    await this.prepareForSave(form);
-    return this.ngZone.runOutsideAngular(async () => {
-      const doc = await this.update(docId);
-      return this._save(form, formDoc, doc);
-    });
+  private buildBinaryAttachmentData(form: string, originalAttachments: Record<string, any>, element: Element) {
+    const xpath = Xpath.getElementTreeXPath(element);
+    const formXpath = xpath.replace(/^\/[^/]+/, `/${form}`);
+    const filename = `${this.USER_BINARY_ATTACHMENT_PREFIX}${formXpath}`;
+    const data = element.textContent;
+    element.textContent = '';
+    return {
+      filename,
+      // Currently do not support loading binary attachment data into edit form. So, keep existing value.
+      attachment: data ? { data, content_type: 'image/png' } : originalAttachments[filename]
+    };
   }
 
-  private async _save(form, formDoc, doc) {
-    const dataString = form.getDataStr({ irrelevant: false });
-    return this.xmlToDocs(doc, formDoc.xml, formDoc.doc.xmlVersion, dataString);
+  private processFormAttachments(
+    form: string,
+    rootData: EnketoRootFormData,
+    originalAttachments: Record<string, any> = {}
+  ) {
+    const binaryAttachments = rootData.binaryTypeElements
+      .map(element => this.buildBinaryAttachmentData(form, originalAttachments, element))
+      .filter(({ attachment }) => attachment)
+      .reduce((acc, { filename, attachment }) => ({ ...acc, [filename]: attachment }), {});
+    const newFileAttachments = FileManager
+      .getCurrentFiles()
+      .map(file => ({
+        name: `${this.USER_FILE_ATTACHMENT_PREFIX}${file.name}`,
+        content_type: file.type,
+        data: new Blob([ file ], { type: file.type })
+      }))
+      .reduce((acc, { name, content_type, data }) => ({ ...acc, [name]: { content_type, data } }), {});
+    const existingFileAttachments = Object
+      .entries(originalAttachments)
+      .filter(([key]) => key.startsWith(this.USER_FILE_ATTACHMENT_PREFIX))
+      // Keep existing file attachments still referenced by a field
+      .filter(([key]) => rootData.findNodeWithTextContent(key.slice(this.USER_FILE_ATTACHMENT_PREFIX.length)))
+      .reduce((acc, [key, attachment]) => ({ ...acc, [key]: attachment }), {});
+
+    const attachments = {
+      ...existingFileAttachments,
+      ...newFileAttachments,
+      ...binaryAttachments
+    };
+    return Object.keys(attachments).length ? attachments : undefined;
   }
 
   unload(enketoForm?: EnketoForm) {
@@ -618,6 +636,7 @@ export interface ContactSummary {
   id: string;
   context: Record<string, any>;
 }
+
 interface ContactSummaryXml {
   id: string;
   xml: Document;
@@ -639,8 +658,6 @@ interface XmlFormContext {
   userContactSummary?: ContactSummary;
 }
 
-export type FormType = 'contact' | 'report' | 'task' | 'training-card';
-
 export interface EnketoFormContext {
   readonly selector: string;
   readonly formConfig: FormConfig;
@@ -653,20 +670,14 @@ export interface EnketoFormContext {
   readonly userContactSummary?: ContactSummary;
 }
 
-export class FormConfig {
-  public readonly repeatPaths: string[];
+export interface EnketoForm {
+  form: Record<string, any>
+  config: FormConfig
+}
 
-  constructor(
-    public readonly doc: Record<string, any>,
-    public type: FormType,
-    xml: string,
-    public readonly html: string,
-    public readonly model: string
-  ) {
-    const xmlDoc = new DOMParser().parseFromString(xml, 'text/xml');
-    this.repeatPaths = Array
-      .from(xmlDoc.querySelectorAll('repeat[nodeset]'))
-      .map(el => el.getAttribute('nodeset')!)
-      .filter(Boolean);
+export class FormValidationError extends Error {
+  constructor(message = 'Form is invalid') {
+    super(message);
+    this.name = 'FormValidationError';
   }
 }
